refactor: replace react with a flask-only version
