+++ .env.example
@@ -1,6 +1,3 @@
-# If running docker locally, set this to http://localhost:3000/api
-REACT_APP_BASE_URL=
-
 # If running docker locally, set this to ff:changeme
 USERS=
