+++ .github/workflows/ci.yml
@@ -29,21 +29,11 @@ jobs:
           username: ${{ secrets.DOCKER_USERNAME }}
           password: ${{ secrets.DOCKER_TOKEN }}
 
-      - name: Build and push frontend
+      - name: Build and push
         uses: docker/build-push-action@v5
         with:
           context: .
-          file: ./Dockerfile.frontend
-          push: true
-          tags: fullfact/raphael-frontend:latest
-          cache-from: type=gha
-          cache-to: type=gha,mode=min
-
-      - name: Build and push backend
-        uses: docker/build-push-action@v5
-        with:
-          context: .
-          file: ./Dockerfile.backend
+          file: ./Dockerfile
           push: true
           tags: fullfact/raphael-backend:latest
           cache-from: type=gha
