--- Dockerfile.frontend
@@ -1,17 +1,0 @@
-FROM node:18.2-alpine AS builder
-
-WORKDIR /app
-COPY src/raphael_frontend_react/package*.json ./
-COPY .env ./.env
-
-RUN npm ci
-
-COPY src/raphael_frontend_react/ ./
-RUN npm run build
-
-FROM nginx:stable-alpine
-COPY --from=builder /app/build /usr/share/nginx/html
-COPY src/raphael_frontend_react/nginx.conf /etc/nginx/conf.d/default.conf
-
-EXPOSE 80
-CMD ["nginx", "-g", "daemon off;"]
