+++ ansible/playbooks/nginx_docker.yaml
@@ -69,20 +69,13 @@
               ssl_certificate {{ ssl_cert_path }};
               ssl_certificate_key {{ ssl_key_path }};
 
-              location ~ ^/api {
+              location / {
                   proxy_pass http://127.0.0.1:3000;
                   proxy_read_timeout 1800;
                   proxy_connect_timeout 1800;
                   proxy_send_timeout 1800;
                   send_timeout 1800;
-              }
-
-              location / {
-                  proxy_pass http://127.0.0.1:4000;
-                  proxy_set_header Host $host;
-                  proxy_set_header X-Real-IP $remote_addr;
-                  proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
-                  proxy_set_header X-Forwarded-Proto $scheme;
+                  proxy_buffering off;
               }
           }
