+++ src/raphael_backend_flask/templates/home.html
@@ -1,0 +1,18 @@
+{% extends 'base.html' %}
+
+{% block content %}
+<h2 class="text-center">Find health misinformation in YouTube videos</h2>
+
+<form method="POST" action="{{ url_for('routes.post_youtube_url') }}" class="form-inline justify-content-center my-4">
+  <div class="form-group mr-2">
+    <input type="text" name="q" placeholder="Enter a YouTube URL" class="form-control" autofocus />
+  </div>
+  <button type="submit" class="btn btn-primary">Analyse</button>
+</form>
+<h3>History</h3>
+<div class="list-group list-group-flush">
+{% for transcript in transcripts %}
+  <a class="list-group-item list-group-item-action" href="{{ url_for('routes.get_video_analysis', video_id=transcript['id']) }}">{{ transcript['metadata']['title'] }}</a>
+{% endfor %}
+</div>
+{% endblock %}
