--- src/raphael_frontend_react/nginx.conf
@@ -1,8 +1,0 @@
-server {
-    listen 80;
-    location / {
-        root /usr/share/nginx/html;
-        index index.html index.htm;
-        try_files $uri $uri/ /index.html;
-    }
-}
