--- src/raphael_frontend_react/src/App.css
@@ -1,12 +1,0 @@
-.App {
-  text-align: left;
-  padding: 24px;
-}
-h6, h4{
-  margin-top: 24px;
-}
-
-.App ul {
-  list-style: none;
-  padding: 0
-}
